Move test to PlannerSuite.
